+++ src/complaints/ComplaintType.js
@@ -3,6 +3,7 @@ import React,{useState} from 'react';
 import ComplaintScreen from './component/PendingComplaint';
 import RecievedComplaint from './component/RecievedComplaint';
 import PendingComplaint from './component/PendingComplaint';
+import { Type } from '../../style/styles';
 
 const ComplaintType = ({navigation}) => {
   const [selectedComplaintType, setSelectedComplaintType] = useState('complaint');
@@ -11,63 +12,31 @@ const ComplaintType = ({navigation}) => {
   };
   return (
     <View>
-      <View style={styles.container}>
+      <View style={Type.container}>
         {/* Text: Pending Complaint */}
         <TouchableOpacity onPress={() => handleButtonPress('complaint')}>
-          <Text style={styles.text}>Pending Complaint</Text>
+          <Text style={Type.text}>Pending Complaint</Text>
         </TouchableOpacity>
 
-        <View style={styles.divider} />
+        <View style={Type.divider} />
        
         {/* Clickable: Received Complaint */}
         <TouchableOpacity onPress={() => handleButtonPress('received')}>
-          <Text style={styles.text}>Received Complaint</Text>
+          <Text style={Type.text}>Received Complaint</Text>
         </TouchableOpacity>
 
-        <View style={styles.divider} />
+        <View style={Type.divider} />
         
         {/* Icon: > */}
-        <Text style={styles.icon}>{'>'}</Text>
+        <Text style={Type.icon}>{'>'}</Text>
       </View>
 
-      <Text style={styles.footerText}>Please Choose The Complaint Type</Text>
+      <Text style={Type.footerText}>Please Choose The Complaint Type</Text>
       {selectedComplaintType === 'complaint' && <PendingComplaint />}
       {selectedComplaintType === 'received' && <RecievedComplaint />}
     </View>
   );
 };
 
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: 'row', // equivalent to 'flex-row'
-    alignItems: 'center', // equivalent to 'items-center'
-    justifyContent: 'space-between', // equivalent to 'justify-between'
-    marginHorizontal: 16, // equivalent to 'mx-4'
-    padding: 16, // equivalent to 'p-4'
-    backgroundColor: '#60a5fa', // equivalent to 'bg-blue-400'
-    borderRadius: 9999, // equivalent to 'rounded-full'
-    marginTop: 32, // equivalent to 'mt-8'
-  },
-  text: {
-    color: 'black', // equivalent to 'text-black'
-    fontSize: 18, // equivalent to 'text-lg'
-  },
-  divider: {
-    borderRightWidth: 1,
-    borderRightColor: 'black', // equivalent to 'border-r border-black'
-    height: 32, // equivalent to 'h-8'
-  },
-  icon: {
-    color: 'white', // equivalent to 'text-white'
-    fontSize: 18, // equivalent to 'text-lg'
-  },
-  footerText: {
-    color: 'black', // equivalent to 'text-black'
-    fontSize: 18, // equivalent to 'text-lg'
-    fontWeight: 'bold', // equivalent to 'font-bold'
-    textAlign: 'center', // equivalent to 'text-center'
-    paddingVertical: 24, // equivalent to 'p-6'
-  },
-});
 
 export default ComplaintType;
