+++ src/complaints/component/InitiatedComplaint.js
@@ -1,19 +1,20 @@
 import { View, Text, StyleSheet, Image } from 'react-native'
 import React from 'react'
+import { ComplaintStages } from '../../../style/styles';
 
 const InitiatedComplaint = ({ ComplaintData }) => {
   return (
-    <View style={styles.container}>
-        <View style={styles.closedIndicator}/>
-        <View style={styles.textContainer}>
-        <Text style={styles.complaintName}>{ComplaintData.complaintName}</Text>
-        <Text style={styles.otherInfo}>{ComplaintData.otherInfo}</Text>
+    <View style={ComplaintStages.container}>
+        <View style={ComplaintStages.initiatedIndicator}/>
+        <View style={ComplaintStages.textContainer}>
+        <Text style={ComplaintStages.complaintName}>{ComplaintData.complaintName}</Text>
+        <Text style={ComplaintStages.otherInfo}>{ComplaintData.otherInfo}</Text>
         </View>
 
-        <View style={styles.imageContainer}>
+        <View style={ComplaintStages.imageContainer}>
       <Image
         source={ComplaintData.imageSource}
-        style={styles.image}
+        style={ComplaintStages.image}
         resizeMode="cover"
       />
       </View>
@@ -22,47 +23,5 @@ const InitiatedComplaint = ({ ComplaintData }) => {
   );
 };
 
-const styles = StyleSheet.create({
-    container: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        backgroundColor: '#cbd5e1',
-        borderRadius: 10,
-        // overflow: 'hidden',
-        width:400,
-        height:200,
-        marginBottom:10,
-      },
-      closedIndicator: {
-        width: 6,
-        backgroundColor: '#BDAB04',
-        height: '100%', // Makes the line stretch to the height of the container
-      },
-      textContainer: {
-        flex: 1,
-        padding: 10,
-        justifyContent: 'center',
-      },
-      complaintName: {
-        fontSize: 20,
-        color: '#333',
-      },
-      otherInfo: {
-        fontSize: 16,
-        color: '#333',
-      },
-      imageContainer: {
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center',
-        padding: 8,
-      },
-      image: {
-        width: '100%',
-        height: '100%',
-        resizeMode: 'contain',
-        borderRadius: 8,
-      },    
-});
 
 export default InitiatedComplaint
