+++ src/complaints/component/PendingComplaint.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import { ScrollView, View, StyleSheet } from 'react-native';
 import ComplaintBox from './ComplaintBox'; // Adjust the path based on your project structure
-import Navbar from '../../navbar/Navbar';
-import ComplaintType from '../ComplaintType';
+import { Pending } from '../../../style/styles';
 import ClosedComplaint from './ClosedComplaint';
 
 const PendingComplaint = () => {
@@ -18,7 +17,7 @@ const PendingComplaint = () => {
     <ScrollView>
         {/* <Navbar /> */}
         {/* <ComplaintType /> */}
-      <View style={styles.container}>
+      <View style={Pending.container}>
         {complaintDataArray.map((ComplaintData, index) => (
           <ComplaintBox key={index} ComplaintData={ComplaintData} />
         ))}
@@ -27,11 +26,6 @@ const PendingComplaint = () => {
   );
 };
 
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-  },
-});
+
 
 export default PendingComplaint;
