+++ src/complaints/component/RecievedComplaint.js
@@ -6,11 +6,14 @@ import ComplaintType from '../ComplaintType';
 import ClosedComplaint from './ClosedComplaint';
 import InProgressComplaint from './InProgressComplaint';
 import InitiatedComplaint from './InitiatedComplaint';
+import SingleComplaint from './SingleComplaint';
+import PendingComplaint from './PendingComplaint';
+import { Pending } from '../../../style/styles';
 
 const RecievedComplaint = ({navigation}) => {
   const complaintDataArray = [
     { imageSource: require('../../../assets/images/leaves.jpg'), complaintName: 'Complaint 1', otherInfo: 'Information of dummy text how about you now how are you ' },
-    { imageSource: require('../../../assets/images/leaves.jpg'), complaintName: 'Complaint 1', otherInfo: 'Information of' },
+    // { imageSource: require('../../../assets/images/leaves.jpg'), complaintName: 'Complaint 1', otherInfo: 'Information of' },
     // { imageSource: require('../../assets/images/leaves.jpg'), complaintName: 'Complaint 1', otherInfo: 'Information of' },
     // { imageSource: require('../../assets/images/leaves.jpg'), complaintName: 'Complaint 1', otherInfo: 'Information of' },
     // ... more data
@@ -18,25 +21,20 @@ const RecievedComplaint = ({navigation}) => {
 
   return (
     <ScrollView>
-        {/* <Navbar /> */}
+        <Navbar />
         {/* <ComplaintType navigation={navigation} /> */}
-      <View style={styles.container}>
+      <View style={Pending.container}>
         {complaintDataArray.map((ComplaintData, index) => (
-        //   <ComplaintBox key={index} ComplaintData={ComplaintData} />
-        //   <ClosedComplaint key={index} ComplaintData={ComplaintData}/>
+          // <ComplaintBox key={index} ComplaintData={ComplaintData} />
+          // <ClosedComplaint key={index} ComplaintData={ComplaintData}/>
         // <InProgressComplaint key={index} ComplaintData={ComplaintData}/>
-        <InitiatedComplaint key={index} ComplaintData={ComplaintData} />
+        // <InitiatedComplaint key={index} ComplaintData={ComplaintData} />
+        <SingleComplaint key={index} ComplaintData={ComplaintData} />
+        // <PendingComplaint key={index} ComplaintData={ComplaintData} />
         ))}
       </View>
     </ScrollView>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-  },
-});
 
 export default RecievedComplaint;
