+++ src/screens/Signin.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { View, Text, TextInput, TouchableOpacity, StyleSheet } from 'react-native';
 import { FIREBASE_AUTH } from '../../Firebase';
-
+import { user } from '../../style/styles';
 
 import { signInWithEmailAndPassword } from 'firebase/auth';
 const Signin = ({navigation}) => {
@@ -27,11 +27,11 @@ const Signin = ({navigation}) => {
   };
 
   return (
-    <View style={styles.container}>
-      <Text style={styles.title}>Sign In</Text>
+    <View style={user.container}>
+      <Text style={user.title}>Sign In</Text>
 
       <TextInput
-        style={styles.input}
+        style={user.input}
         placeholder="Email"
         value={email}
         onChangeText={setEmail}
@@ -40,51 +40,20 @@ const Signin = ({navigation}) => {
       />
 
       <TextInput
-        style={styles.input}
+        style={user.input}
         placeholder="Password"
         value={password}
         onChangeText={setPassword}
         secureTextEntry
       />
 
-      <TouchableOpacity style={styles.button} onPress={handleSignin}>
-        <Text style={styles.buttonText}>Sign In</Text>
+      <TouchableOpacity style={user.button} onPress={handleSignin}>
+        <Text style={user.buttonText}>Sign In</Text>
         {/* // adnan530@gmail.com addi3214 & adnan0@gmail.com kamranbhai addi12@gmail.com 12345612*/}
       </TouchableOpacity> 
     </View>
   );
 };
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: 20,
-  },
-  title: {
-    fontSize: 24,
-    marginBottom: 20,
-  },
-  input: {
-    width: '100%',
-    borderWidth: 1,
-    borderColor: 'gray',
-    padding: 10,
-    borderRadius: 5,
-    marginBottom: 10,
-  },
-  button: {
-    backgroundColor: 'blue',
-    padding: 15,
-    borderRadius: 5,
-    width: '100%',
-    alignItems: 'center',
-  },
-  buttonText: {
-    color: 'white',
-    fontWeight: 'bold',
-  },
-});
 
 export default Signin;
